The mentor create page was added and the delete logic was improved
