+++ src/app/mentors/create/page.tsx
@@ -1,0 +1,11 @@
+import MentorsForm from "@/components/MentorForm";
+import Typography from "@/components/Typography/typography";
+
+export default function CreateMentorForm() {
+    return (
+        <>
+            <Typography tag="h1">Crear Mentor</Typography>
+            <MentorsForm></MentorsForm>
+        </>
+    );
+}
